Add test_default_random
